move parallel util
