+++ cpp/open3d/core/kernel/CPULauncher.h
@@ -31,8 +31,8 @@
 
 #include "open3d/core/AdvancedIndexing.h"
 #include "open3d/core/Indexer.h"
-#include "open3d/core/ParallelUtil.h"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/kernel/ParallelUtil.h"
 #include "open3d/utility/Console.h"
 
 namespace open3d {
@@ -103,7 +103,7 @@ public:
                     "single-output reduction ops.");
         }
         int64_t num_workloads = indexer.NumWorkloads();
-        int64_t num_threads = parallel_util::GetMaxThreads();
+        int64_t num_threads = GetMaxThreads();
         int64_t workload_per_thread =
                 (num_workloads + num_threads - 1) / num_threads;
         std::vector<scalar_t> thread_results(num_threads, identity);
@@ -132,7 +132,7 @@ public:
         // Prefers outer dimension >= num_threads.
         const int64_t* indexer_shape = indexer.GetMasterShape();
         const int64_t num_dims = indexer.NumDims();
-        int64_t num_threads = parallel_util::GetMaxThreads();
+        int64_t num_threads = GetMaxThreads();
 
         // Init best_dim as the outer-most non-reduction dim.
         int64_t best_dim = num_dims - 1;
